Rename `CanonReflect` and reformulate the problem
By community consensus, `CanonReflect` is now `Reflect` and "old"
`Reflect` is now `PartialReflect`.

Since we "keep" `Reflect`, I reformulate the change to conceptually fit
with the current reference frame.

I also explain in more details the concepts `Reflect` represent and
make clear what the true underlying source of the awkwardness of
`Reflect` is.